+++ pytest_django/plugin.py
@@ -57,6 +57,7 @@ if TYPE_CHECKING:
     from typing import Any, NoReturn
 
     import django
+    import django.apps.registry
 
 
 SETTINGS_MODULE_ENV = "DJANGO_SETTINGS_MODULE"
@@ -293,6 +294,10 @@ def pytest_load_initial_conftests(
         "the `urls` attribute of Django's `TestCase` objects.  *modstr* is "
         "a string specifying the module of a URL config, e.g. "
         '"my_app.test_urls".',
+    )
+    early_config.addinivalue_line(
+        "markers",
+        "django_isolate_apps(*app_labels): isolate Django's app registry for this test.",
     )
     early_config.addinivalue_line(
         "markers",
@@ -669,6 +674,39 @@ def _django_set_urlconf(request: pytest.FixtureRequest) -> Generator[None, None,
         set_urlconf(None)
 
 
+@pytest.fixture(autouse=True)
+def _django_isolate_apps(
+    request: pytest.FixtureRequest,
+) -> Generator[django.apps.registry.Apps, None, None]:
+    """Apply the @pytest.mark.django_isolate_apps marker if present, internal to pytest-django."""
+    marker: pytest.Mark | None = request.node.get_closest_marker("django_isolate_apps")
+    if not marker:
+        yield None
+        return
+
+    skip_if_no_django()
+
+    from django.test.utils import isolate_apps
+
+    app_labels = validate_django_isolate_apps(marker)
+
+    with isolate_apps(*app_labels) as apps:
+        yield apps
+
+
+@pytest.fixture
+def django_isolated_apps(
+    _django_isolate_apps: django.apps.registry.Apps | None,
+) -> django.apps.registry.Apps:
+    """Access the isolated Apps registry instance for tests marked with
+    @pytest.mark.django_isolate_apps(...)."""
+    if _django_isolate_apps is None:
+        raise pytest.UsageError(
+            "The django_isolated_apps fixture requires @pytest.mark.django_isolate_apps([...])."
+        )
+    return _django_isolate_apps
+
+
 @pytest.fixture(autouse=True, scope="session")
 def _fail_for_invalid_template_variable() -> Generator[None, None, None]:
     """Fixture that fails for invalid variables in templates.
@@ -885,5 +923,16 @@ def validate_urls(marker: pytest.Mark) -> list[str]:
 
     def apifun(urls: list[str]) -> list[str]:
         return urls
+
+    return apifun(*marker.args, **marker.kwargs)
+
+
+def validate_django_isolate_apps(marker: pytest.Mark) -> tuple[str, ...]:
+    """Validate the django_isolate_apps marker."""
+
+    def apifun(*app_labels: str) -> tuple[str, ...]:
+        if not app_labels:
+            raise ValueError("@pytest.mark.django_isolate_apps requires at least one app label")
+        return app_labels
 
     return apifun(*marker.args, **marker.kwargs)