+++ pytest_django_test/settings_mysql.py
@@ -45,6 +45,7 @@ DATABASES = {
         "TEST": {
             "CHARSET": "utf8mb4",
             "COLLATION": "utf8mb4_unicode_ci",
+            "DEPENDENCIES": [],
         },
     },
 }